Naive Paging Plugin Implemented
Paging plugin allows the enclave to use back-up memory as a swap space.
When there is not sufficient free memory, the paging plugin will evict
one of the valid user page to free a frame, and allocate the frame.
Page fault handler is hooked when the plugin initializes such that the
evicted page can be loaded.

Currently, the paging plugin uses random eviction policy, where it picks
a random counter and search for the n-th user page by traversing the page table.

Paging plugin can be compiled with `-DUSE_PAGING` flag, but it also
requires `-DUSE_FREEMEM`.