Update travis to use SDK's new CMake build system, fix SDK revision
Add SDK_MIN_REVISION file for the benefit of Travis and the user, use
that file to checkout a particular SDK revision.

This allows the current repository to use the SDK version for keystone's
dev branch, which uses a different (CMake-based) build system. Updates
the travis build script to use that.